Keep fetching integration tests easier to extend
The fetching integration suite repeated production-context, request-context, adapter, and typed-prop helpers inline, so this extracts those shared test utilities into a sibling helper module without changing data-fetching assertions.

Constraint: Reclaimed from stale team state; keep the refactor narrow and test-only while preserving the existing fetching integration coverage exactly
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Convert the suite into broader table-driven helpers | would widen the change and make the cache/revalidation scenarios harder to read during the wave
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep repeated fetching test utilities in the helper module unless a future case needs bespoke inline setup for a specific data path
Tested: NODE_ENV=production deno test --no-check --allow-all --unstable-worker-options tests/integration/data/fetching/fetching.test.ts; deno fmt --check tests/integration/data/fetching/fetching.test.ts tests/integration/data/fetching/fetching.test-helpers.ts; deno lint tests/integration/data/fetching/fetching.test.ts tests/integration/data/fetching/fetching.test-helpers.ts; git diff --check
Not-tested: Full veryfront-code repo suite
